# backend/internal/shared/apidocs/openapi.yaml
openapi: 3.0.0
info:
    title: API Docs
    version: 1.0.0
paths:
    /api/auth/invite:
        post:
            summary: Invite User
            tags:
                - Auth
            parameters:
                - name: Content-Type
                  in: header
                  required: true
                  example: application/json
                - name: Authorization
                  in: header
                  required: true
                  example: Bearer {{AUTH_TOKEN}}
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                department:
                                    type: string
                                email:
                                    type: string
                                first_name:
                                    type: string
                                last_name:
                                    type: string
                                phone:
                                    type: string
                                role:
                                    type: string
                                role_at_org:
                                    type: string
                        example: |4-
                              {
                                "first_name": "Law",
                                "last_name": "Kwis",
                                "email": "samuelfadunsin@gmail.com",
                                "role": "admin",
                                "role_at_org": "Backend Engineer",
                                "department": "Engineering",
                                "phone": "+2347046129850"
                              }
            responses:
                "200":
                    description: OK
    /api/auth/magic-link:
        post:
            summary: Request Magic Link
            tags:
                - Auth
            parameters:
                - name: Content-Type
                  in: header
                  required: true
                  example: application/json
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                        example: |4-
                              {
                                  "email": "joshuaajagbe96@gmail.com"
                              }
            responses:
                "200":
                    description: OK
    /api/auth/magic-link/verify:
        post:
            summary: Verify Magic Link
            tags:
                - Auth
            parameters:
                - name: Content-Type
                  in: header
                  required: true
                  example: application/json
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                token:
                                    type: string
                        example: |4-
                              {
                                "token": "c9bbe34f9ab4537e3956298a61fbd328634698ffe9029b2034ee89f68b252ed8"
                              }
            responses:
                "200":
                    description: OK
    /api/chat/conversations:
        get:
            summary: List Active Conversations
            tags:
                - Messaging
            responses:
                "200":
                    description: OK
    /api/chat/create:
        post:
            summary: Initiate Conversation
            tags:
                - Messaging
            parameters:
                - name: Content-Type
                  in: header
                  required: true
                  example: application/json
                - name: Authorization
                  in: header
                  required: true
                  example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxNTMxYzJiMS01ZjU0LTQyM2MtOGM0MC03Y2U3OGNlMzA4Y2EiLCJlbWFpbCI6Impvc2h1YWFqYWdiZTk2QGdtYWlsLmNvbSIsInJvbGUiOiJhZG1pbiIsInRva2VuX3R5cGUiOiJhY2Nlc3MiLCJleHAiOjE3NzIyMDQyODYsImlhdCI6MTc3MTU5OTQ4NiwianRpIjoiOTIwOTI5NDEtM2Q3NC00OTIxLWE3M2EtYWM3YjgzMjc5NTBlIn0.q-inoYLk6pAw4Kn6foklpKhaDbztN3qmZjev_0JGjaU
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: string
                        example: |4-
                              {
                                "channel_type": "dm",
                              //   "channel_name": "playground",
                                "channel_description": "This is a playground group for testing purposes.",
                                "channel_members" : [{"email" : "samuelfadunsin@gmail.com", "role" : "admin" }]
                              }
            responses:
                "200":
                    description: OK
    /api/chat/messages?limit=10&conversation_id=cccdfcaa-b0bc-47c3-9edd-899c7acdc519:
        get:
            summary: Get Conversation Messages
            tags:
                - Messaging
            parameters:
                - name: limit
                  in: query
                  required: true
                  example: "10"
                - name: conversation_id
                  in: query
                  required: true
                  example: cccdfcaa-b0bc-47c3-9edd-899c7acdc519
                - name: Content-Type
                  in: header
                  required: true
                  example: text/plain
                - name: Authorization
                  in: header
                  required: true
                  example: Bearer {{AUTH_TOKEN}}
            requestBody:
                required: true
                content:
                    text/plain:
                        schema:
                            type: string
                        example: |4-
                              {
                                  "token": "e54c7462488d0f3c8e8e42aa1da9c75c0602ba32ac5ad81880cfb0cb531d37ea"
                              }
            responses:
                "200":
                    description: OK
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
security:
    - bearerAuth: []
    - bearerAuth: []
    - bearerAuth: []
    - bearerAuth: []